# services/gateway/src/openapi.yaml
openapi: 3.0.3
info:
  title: Dental Clinic Appointment API
  version: "1.0.0"
  description: >
    Public API surface for the dental appointment booking system, served
    behind the gateway. Internally this fans out to independent
    microservices (patient, dentist, appointment, reminder, notification,
    report) — but every route below is reached the same way it always was,
    as a single relative `/api/*` origin.
servers:
  - url: /api
    description: Through the gateway (relative — works from any host it's deployed behind)
  - url: http://localhost:4000/api
    description: Local dev
...
paths:
...
  /patients:
    get:
      tags: [Patients]
      summary: List / search patients
      parameters:
        - name: q
          in: query
          description: Free-text search over name, email, phone
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Patient" }
    post:
      tags: [Patients]
      summary: Create a patient
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PatientInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Patient" }
        "400":
          description: name and email are required
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /patients/{id}:
    get:
      tags: [Patients]
      summary: Get a patient, including their appointment history
      parameters:
        - $ref: "#/components/parameters/PatientId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PatientWithHistory" }
        "404":
          description: patient not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    put:
      tags: [Patients]
      summary: Update a patient
      parameters:
        - $ref: "#/components/parameters/PatientId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PatientInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Patient" }
        "400":
          description: name and email are required
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: patient not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /dentists:
    get:
      tags: [Dentists]
      summary: List dentists
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Dentist" }
    post:
      tags: [Dentists]
      summary: Create a dentist
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DentistInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Dentist" }
        "400":
          description: Invalid name / slot_minutes / working hours
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /dentists/{id}/slots:
    get:
      tags: [Dentists]
      summary: Bookable slots for a dentist on a given day
      parameters:
        - $ref: "#/components/parameters/DentistId"
        - name: date
          in: query
          required: true
          description: YYYY-MM-DD, interpreted in server-local time
          schema: { type: string, format: date }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  dentist: { $ref: "#/components/schemas/Dentist" }
                  date: { type: string, format: date }
                  slots:
                    type: array
                    items: { $ref: "#/components/schemas/Slot" }
        "400":
          description: date=YYYY-MM-DD is required
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: dentist not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503":
          description: could not check dentist availability, try again
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /appointments:
    get:
      tags: [Appointments]
      summary: List appointments
      parameters:
        - name: status
          in: query
          description: Comma-separated statuses
          schema: { type: string }
        - name: from
          in: query
          schema: { type: string, format: date-time }
        - name: to
          in: query
          schema: { type: string, format: date-time }
        - name: dentistId
          in: query
          schema: { type: integer }
        - name: patientId
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Appointment" }
    post:
      tags: [Appointments]
      summary: Book an appointment
      description: >
        Validates patient_id/dentist_id against patient-/dentist-service
        (fails closed with 503 if either is unreachable), checks for a
        conflicting slot (409), schedules reminders, and fires a
        confirmation email.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BookingInput" }
      responses:
        "201":
          description: Booked
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Appointment" }
        "400":
          description: Invalid patient_id/dentist_id/time, or time in the past
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409":
          description: that slot is already booked (double-booking prevented)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503":
          description: unable to validate patient/dentist right now, please retry
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
components:
  parameters:
    PatientId:
      name: id
      in: path
      required: true
      schema: { type: integer }
    DentistId:
      name: id
      in: path
      required: true
      schema: { type: integer }
...
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
...
    Patient:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string, format: email }
        phone: { type: string, nullable: true }
        dob: { type: string, format: date, nullable: true }
        notes: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
    PatientWithHistory:
      allOf:
        - $ref: "#/components/schemas/Patient"
        - type: object
          properties:
            appointments:
              type: array
              items: { $ref: "#/components/schemas/Appointment" }
    PatientInput:
      type: object
      required: [name, email]
      properties:
        name: { type: string }
        email: { type: string, format: email }
        phone: { type: string }
        dob: { type: string, format: date }
        notes: { type: string }
    Dentist:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        specialty: { type: string, nullable: true }
        email: { type: string, nullable: true }
        work_start: { type: integer, description: Hour of day, 24h, example: 9 }
        work_end: { type: integer, description: Hour of day, 24h, example: 17 }
        slot_minutes: { type: integer, example: 30 }
    DentistInput:
      type: object
      required: [name]
      properties:
        name: { type: string }
        specialty: { type: string }
        email: { type: string }
        work_start: { type: integer, default: 9 }
        work_end: { type: integer, default: 17 }
        slot_minutes: { type: integer, default: 30, description: Must be > 0 }
    Slot:
      type: object
      properties:
        start: { type: string, format: date-time }
        end: { type: string, format: date-time }
        available: { type: boolean }
    Appointment:
      type: object
      properties:
        id: { type: integer }
        patient_id: { type: integer }
        dentist_id: { type: integer }
        start_time: { type: string, format: date-time }
        end_time: { type: string, format: date-time }
        status: { type: string, enum: [booked, completed, cancelled, no_show] }
        reason: { type: string, nullable: true }
        notes: { type: string, nullable: true }
        follow_up_of: { type: integer, nullable: true }
        cancel_reason: { type: string, nullable: true }
        cancelled_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
        patient_name: { type: string, nullable: true, description: Composed from patient-service }
        patient_email: { type: string, nullable: true, description: Composed from patient-service }
        patient_phone: { type: string, nullable: true, description: Composed from patient-service }
        dentist_name: { type: string, nullable: true, description: Composed from dentist-service }
    BookingInput:
      type: object
      required: [patient_id, dentist_id, start]
      properties:
        patient_id: { type: integer }
        dentist_id: { type: integer }
        start: { type: string, format: date-time }
        end: { type: string, format: date-time, description: "Optional — defaults from duration_minutes / the dentist's slot length" }
        duration_minutes: { type: integer }
        reason: { type: string }
        notes: { type: string }
...